requirements.txt
# Pinned runtime dependencies.
#
# Source of truth: requirements.in

aiosqlite==0.20.0
discord.py==2.4.0
praw==7.7.1
python-dotenv==1.0.1
